Workaround touch event bugs on iOS
On iOS, if an element is created from a document that is not window.document, touch event listeners will not work
correctly, and may crash the browser if the element is removed.

Therefore, for iOS we implement only one touch listener on document, and use target finding on touchstart.

These broken touch handlers are very easy to make with the Polymer declarative gesture system, but this is a bug for
Safari to fix.

Fixes #56 Work around safari crasher
Fixes #33 Work around safari touchevent listener bug